% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrecipLZ.R
\name{getQ}
\alias{getQ}
\title{Get the observations}
\usage{
getQ(method = NULL, path = NULL, filename = NULL)
}
\arguments{
\item{method}{method for processing observations, so far only "processedNVE"}

\item{path}{Directory where to get the files}

\item{filename}{Filename (path included) of the file to be read}
}
\description{
Get and process, if needed, the observations
}
\examples{
\dontrun{
getQ()
}
}
\keyword{Q}

